pyproject.toml
[tool.poetry]
name = "llm-hackathon"
version = "0.1.0"
description = ""
authors = ["Christian Lehre <christian.lehre@akerbp.com>"]
readme = "README.md"
packages = [{include = "llm_hackathon"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
streamlit-chat = "^0.1.1"
langchain = "^0.0.264"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
sentence-transformers = "^2.2.2"
pypdf = "^3.15.1"
faiss-cpu = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
